requirements.txt
mock==4.0.2
numpy==1.16.3
pandas==1.1.0
pyyaml==5.3.1
requests==2.24.0
scipy==1.5.2
tensorflow==2.3.0
xlrd==1.2.0